# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HUBBIES PUPPIES API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://dev.hubbies-puppies.fr",
            "description": "HUBBIES PUPPIES DEV API"
        },
        {
            "url": "http://staging.hubbies-puppies.fr",
            "description": "HUBBIES PUPPIES STAGING PI"
        },
        {
            "url": "http://preprod.hubbies-puppies.fr",
            "description": "HUBBIES PUPPIES PREPROD API"
        }
    ],
    "paths": {
...
        "/animals/{animalId}": {
            "get": {
                "tags": [
                    "Animal"
                ],
                "summary": "Get a single Animal from uuid",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/animalId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Animal successfully edited",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "$ref": "#/components/schemas/Animal/properties/uuid"
                                        },
                                        "name": {
                                            "$ref": "#/components/schemas/Animal/properties/name"
                                        },
                                        "gender": {
                                            "$ref": "#/components/schemas/Animal/properties/gender"
                                        },
                                        "age": {
                                            "$ref": "#/components/schemas/Animal/properties/age"
                                        },
                                        "description": {
                                            "$ref": "#/components/schemas/Animal/properties/description"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/animal_404"
                    },
                    "500": {
                        "$ref": "#/components/responses/error_500"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Animal"
                ],
                "summary": "Edit an Animal",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/animalId"
                    }
                ],
                "requestBody": {
                    "description": "Animal that will be edited",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "$ref": "#/components/schemas/Animal/properties/name"
                                    },
                                    "gender": {
                                        "$ref": "#/components/schemas/Animal/properties/gender"
                                    },
                                    "age": {
                                        "$ref": "#/components/schemas/Animal/properties/age"
                                    },
                                    "description": {
                                        "$ref": "#/components/schemas/Animal/properties/description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Animal successfully edited",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "$ref": "#/components/schemas/Animal/properties/uuid"
                                        },
                                        "name": {
                                            "$ref": "#/components/schemas/Animal/properties/name"
                                        },
                                        "gender": {
                                            "$ref": "#/components/schemas/Animal/properties/gender"
                                        },
                                        "age": {
                                            "$ref": "#/components/schemas/Animal/properties/age"
                                        },
                                        "description": {
                                            "$ref": "#/components/schemas/Animal/properties/description"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/animal_404"
                    },
                    "500": {
                        "$ref": "#/components/responses/error_500"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Animal"
                ],
                "summary": "Delete an Animal",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/animalId"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "no-content"
                    },
                    "404": {
                        "$ref": "#/components/responses/animal_404"
                    },
                    "500": {
                        "$ref": "#/components/responses/error_500"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "responses": {
            "animal_404": {
                "description": "Animal not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "code": {
                                    "description": "Access those value with : #/components/responses/{response}",
                                    "example": "404"
                                },
                                "message": {
                                    "description": "Access those value with : #/components/responses/{response}",
                                    "example": "animalNotFound"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "error_500": {
                "description": "Error: Internal Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "code": {
                                    "description": "Access those value with : #/components/responses/{response}",
                                    "example": "500"
                                },
                                "message": {
                                    "description": "Access those value with : #/components/responses/{response}",
                                    "example": "systemError"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "parameters": {
            "animalId": {
                "name": "animalId",
                "in": "path",
                "description": "uuid of an animal",
                "required": true,
                "schema": {
                    "type": "string"
                },
                "example": "bd6bc279-2a0c-40cc-ae94-cb6c1697cf64"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}